🥷 divide package
