+++ main.go
@@ -5,12 +5,12 @@ import (
 	"fmt"
 	"io/ioutil"
 	"log"
-	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
 	"runtime"
 	"runtime/debug"
+
+	"github.com/Omochice/deepl-translate-cli/deepl"
 
 	"github.com/mattn/go-isatty"
 	"github.com/urfave/cli/v2"
@@ -34,82 +34,11 @@ func getVersion() string {
 	return i.Main.Version
 }
 
-type DeepL interface {
-	Translate(Text string, setting Setting) ([]string, error)
-}
-
-type DeepLClient struct {
-	Endpoint string
-	AuthKey  string
-}
-
-func (c *DeepLClient) Translate(text string, sourceLang string, targetLang string) ([]string, error) {
-	params := url.Values{}
-	params.Add("auth_key", c.AuthKey)
-	params.Add("source_lang", sourceLang)
-	params.Add("target_lang", targetLang)
-	params.Add("text", text)
-	resp, err := http.PostForm(c.Endpoint, params)
-
-	if err := ValidateResponse(resp); err != nil {
-		return []string{}, err
-	}
-	parsed, err := ParseResponse(resp)
-	if err != nil {
-		return []string{}, err
-	}
-	r := []string{}
-	for _, translated := range parsed.Translations {
-		r = append(r, translated.Text)
-	}
-	return r, nil
-}
-
-var KnownErrors = map[int]string{
-	400: "Bad request. Please check error message and your parameters.",
-	403: "Authorization failed. Please supply a valid auth_key parameter.",
-	404: "The requested resource could not be found.",
-	413: "The request size exceeds the limit.",
-	414: "The request URL is too long. You can avoid this error by using a POST request instead of a GET request, and sending the parameters in the HTTP body.",
-	429: "Too many requests. Please wait and resend your request.",
-	456: "Quota exceeded. The character limit has been reached.",
-	503: "Resource currently unavailable. Try again later.",
-	529: "Too many requests. Please wait and resend your request.",
-} // this from https://www.deepl.com/docs-api/accessing-the-api/error-handling/
-
-func ValidateResponse(resp *http.Response) error {
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		var data map[string]interface{}
-		baseErrorText := fmt.Sprintf("Invalid response [%d %s]",
-			resp.StatusCode,
-			http.StatusText(resp.StatusCode))
-		if t, ok := KnownErrors[resp.StatusCode]; ok {
-			baseErrorText += fmt.Sprintf(" %s", t)
-		}
-		e := json.NewDecoder(resp.Body).Decode(&data)
-		if e != nil {
-			return fmt.Errorf("%s", baseErrorText)
-		} else {
-			return fmt.Errorf("%s, %s", baseErrorText, data["message"])
-		}
-	}
-	return nil
-}
-
 type Setting struct {
 	AuthKey    string `json:"-"`
 	SourceLang string `json:"source_lang"`
 	TargetLang string `json:"target_lang"`
 	IsPro      bool   `json:"-"`
-}
-
-type DeepLResponse struct {
-	Translations []Translated
-}
-
-type Translated struct {
-	DetectedSourceLaguage string `json:"detected_source_language"`
-	Text                  string `json:"text"`
 }
 
 func LoadSettings(setting Setting, automake bool) (Setting, error) {
@@ -172,29 +101,6 @@ func InitializeConfigFile(ConfigPath string) error {
 
 	out.Write(([]byte)(decoded))
 	return nil
-}
-
-func getUrl(IsPro bool) string {
-	if IsPro {
-		return "https://api.deepl.com/v2/translate"
-	}
-	return "https://api-free.deepl.com/v2/translate"
-}
-
-func ParseResponse(resp *http.Response) (DeepLResponse, error) {
-	var responseJson DeepLResponse
-	body, err := ioutil.ReadAll(resp.Body)
-
-	if err != nil {
-		err := fmt.Errorf("%s (occurred while parse response)", err.Error())
-		return responseJson, err
-	}
-	err = json.Unmarshal(body, &responseJson)
-	if err != nil {
-		err := fmt.Errorf("%s (occurred while parse response)", err.Error())
-		return responseJson, err
-	}
-	return responseJson, err
 }
 
 func main() {
@@ -274,8 +180,8 @@ func main() {
 				rawSentense = string(b)
 			}
 
-			client := DeepLClient{
-				Endpoint: getUrl(c.Bool("pro")),
+			client := deepl.DeepLClient{
+				Endpoint: deepl.GetEndpoint(c.Bool("pro")),
 				AuthKey:  setting.AuthKey,
 			}
 			translateds, err := client.Translate(rawSentense, setting.SourceLang, setting.TargetLang)
